requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
firebase-admin
replicate
httpx
python-dotenv
google-generativeai
pydantic[email]